# docs/openapi-v0.5.1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Olist Decision Backend API",
    "version": "0.5.1"
  },
  "paths": {
...
    "/api/v1/logs/errors": {
      "get": {
        "tags": [
          "Logs"
        ],
        "summary": "List Log Errors",
        "operationId": "list_log_errors_api_v1_logs_errors_get",
        "parameters": [
          {
            "name": "request_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "error_code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Error Code"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorLogListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/logs/audit": {
      "get": {
        "tags": [
          "Logs"
        ],
        "summary": "List Log Audit",
        "operationId": "list_log_audit_api_v1_logs_audit_get",
        "parameters": [
          {
            "name": "outbox_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Outbox Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/logs/recent": {
      "get": {
        "tags": [
          "Logs"
        ],
        "summary": "List Log Recent",
        "operationId": "list_log_recent_api_v1_logs_recent_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentLogListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditLogEntry": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "default": ""
          },
          "outbox_id": {
            "type": "string",
            "title": "Outbox Id",
            "default": ""
          },
          "target_channel": {
            "type": "string",
            "title": "Target Channel",
            "default": ""
          },
          "adapter_name": {
            "type": "string",
            "title": "Adapter Name",
            "default": ""
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "api"
          }
        },
        "type": "object",
        "title": "AuditLogEntry"
      },
      "AuditLogListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditLogEntry"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "AuditLogListResponse"
      },
...
      "ErrorLogEntry": {
        "properties": {
          "ts": {
            "type": "string",
            "title": "Ts"
          },
          "level": {
            "type": "string",
            "title": "Level",
            "default": "ERROR"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          },
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "default": ""
          },
          "diagnosis": {
            "type": "string",
            "title": "Diagnosis",
            "default": ""
          },
          "suggested_action": {
            "type": "string",
            "title": "Suggested Action",
            "default": ""
          },
          "actor": {
            "type": "string",
            "title": "Actor",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "ts"
        ],
        "title": "ErrorLogEntry"
      },
      "ErrorLogListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ErrorLogEntry"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ErrorLogListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecentLogEntry": {
        "properties": {
          "ts": {
            "type": "string",
            "title": "Ts"
          },
          "level": {
            "type": "string",
            "title": "Level",
            "default": ""
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          },
          "method": {
            "type": "string",
            "title": "Method",
            "default": ""
          },
          "path": {
            "type": "string",
            "title": "Path",
            "default": ""
          },
          "actor": {
            "type": "string",
            "title": "Actor",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "ts"
        ],
        "title": "RecentLogEntry"
      },
      "RecentLogListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecentLogEntry"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "RecentLogListResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}